🎨 Add a SwanClient class
